+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachang
 - check-jenkins-job-status.rb: if a job is in progress, use the result of the last completed build instead
 - Update to rubocop 0.40 and cleanup
 - Relax `sensu-plugin` dependency
+- Added authentication options to check-jenkins-build-time.rb and check-jenkins-job-status.rb
 
 ## [0.1.0] - 2015-12-14
 ### Added